+++ packages/elements/src/Screen.tsx
@@ -6,7 +6,7 @@ import {
   useTheme,
 } from '@react-navigation/native';
 import * as React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { Platform, StyleSheet, View, type ViewProps } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
 import { Container, type Props as ContainerProps } from './Container';
@@ -24,6 +24,8 @@ type Props = {
   headerShown?: boolean | undefined;
   headerStatusBarHeight?: number | undefined;
   headerTransparent?: boolean | undefined;
+  pageOverflowEnabled?: boolean | undefined;
+  contentStyle?: ViewProps['style'] | undefined;
   style?: ContainerProps['style'] | undefined;
   children: React.ReactNode;
 };
@@ -46,6 +48,8 @@ export function Screen(props: Props) {
     route,
     children,
     style,
+    contentStyle,
+    pageOverflowEnabled = false,
   } = props;
 
   const defaultHeaderHeight = useFrameSize((size) =>
@@ -55,6 +59,43 @@ export function Screen(props: Props) {
       topInset: headerStatusBarHeight,
     })
   );
+
+  // If the container fills the body, we consider it to be a "page" that can overflow the screen
+  // So we adjust the style accordingly to let the content take more space if needed
+  // This lets the document.body handle scrolling of the content
+  // This is necessary for mobile browsers to be able to hide the address bar on scroll
+  const [fill, setFill] = React.useState(false);
+
+  const onRef = React.useCallback(
+    (node: HTMLDivElement | React.ComponentRef<typeof View> | null) => {
+      if (
+        Platform.OS !== 'web' ||
+        !pageOverflowEnabled ||
+        !(node instanceof HTMLElement)
+      ) {
+        return;
+      }
+
+      const updateFill = () => {
+        setFill(
+          node.clientWidth === document.body.clientWidth &&
+            node.clientHeight === document.body.clientHeight
+        );
+      };
+
+      updateFill();
+
+      const observer = new ResizeObserver(updateFill);
+
+      observer.observe(node);
+      observer.observe(document.body);
+
+      return () => observer.disconnect();
+    },
+    [pageOverflowEnabled]
+  );
+
+  const page = pageOverflowEnabled && focused && fill;
 
   const headerRef = React.useRef<React.ComponentRef<typeof View>>(null);
 
@@ -68,6 +109,7 @@ export function Screen(props: Props) {
 
   return (
     <Container
+      ref={onRef}
       inert={!focused}
       style={{
         ...styles.container,
@@ -107,7 +149,7 @@ export function Screen(props: Props) {
           </View>
         </NavigationProvider>
       ) : null}
-      <View style={styles.content}>
+      <View style={[page ? styles.page : styles.content, contentStyle]}>
         <HeaderShownContext.Provider
           value={isParentHeaderShown || headerShown !== false}
         >
@@ -128,6 +170,9 @@ const styles = StyleSheet.create({
   },
   content: {
     flex: 1,
+  },
+  page: {
+    flexGrow: 1,
   },
   header: {
     zIndex: 1,